# src/main/java/edu/upc/upcschool/mortgage/controllers/MortgageController.java
package edu.upc.upcschool.mortgage.controllers;

import edu.upc.upcschool.mortgage.models.Mortgage;
import edu.upc.upcschool.mortgage.models.Bank;
import edu.upc.upcschool.mortgage.models.User;
import edu.upc.upcschool.mortgage.security.ApiKeyAuth;
import org.springframework.security.core.context.SecurityContextHolder;
import edu.upc.upcschool.mortgage.repositories.BankRepository;
import edu.upc.upcschool.mortgage.repositories.MortgageRepository;
import edu.upc.upcschool.mortgage.services.MortgageSimulationService;
import org.springframework.http.ResponseEntity;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.HttpHeaders;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.util.UriComponentsBuilder;

import java.net.URI;
import java.util.List;
import java.util.Optional;

import static org.springframework.http.ResponseEntity.ok;

@RestController
@RequestMapping({ "/banks/{bankId}/mortgages" })
public class MortgageController {

    private final MortgageRepository mortgageRepository;
    private final BankRepository bankRepository;
    private final MortgageSimulationService mortgageSimulationService;

    /**
     * Initializing Mortgage Controller via class Constructor, preventing null
     * access
     * before creation at execution.
     */
    public MortgageController(MortgageRepository mortgageRepository,
            BankRepository bankRepository,
            MortgageSimulationService mortgageSimulationService) {
        this.mortgageRepository = mortgageRepository;
        this.bankRepository = bankRepository;
        this.mortgageSimulationService = mortgageSimulationService;
    }

    @GetMapping
    public ResponseEntity<List<Mortgage>> findAll(@PathVariable Integer bankId) {
        if (!bankRepository.existsById(bankId)) {
            return ResponseEntity.notFound().build();
        }
        return ResponseEntity.ok(mortgageRepository.findByBankId(bankId));
    }

    @GetMapping("/{id}")
    public ResponseEntity<Mortgage> findById(@PathVariable Integer bankId,
            @PathVariable Integer id) {
        if (!bankRepository.existsById(bankId)) {
            return ResponseEntity.notFound().build();
        }
        return mortgageRepository.findByIdAndBankId(id, bankId)
                .map(ResponseEntity::ok)
                .orElse(ResponseEntity.notFound().build());
    }

    @PostMapping
    public ResponseEntity<Mortgage> create(@PathVariable Integer bankId,
            @RequestBody Mortgage newMortgage,
            UriComponentsBuilder ucb) {

        ApiKeyAuth auth = (ApiKeyAuth) SecurityContextHolder.getContext().getAuthentication();
        User user = (User) auth.getPrincipal();

        Optional<Bank> bankOptional = bankRepository.findById(bankId);
        if (bankOptional.isEmpty()) {
            return ResponseEntity.notFound().build();
        }
        if (!bankOptional.get().ownerId().equals(user.id())) {
            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
        }

        Mortgage toSave = new Mortgage(null, bankId, newMortgage.name(), newMortgage.type(), newMortgage.description(),
                newMortgage.TAE());
        Mortgage saved = mortgageRepository.save(toSave);

        URI location = ucb
                .path("/banks/{bankId}/mortgages")
                .buildAndExpand(bankId)
                .toUri();
        return ResponseEntity.created(location).body(saved);
    }

    @PutMapping("/{id}")
    public ResponseEntity<Mortgage> update(@PathVariable Integer bankId,
            @PathVariable Integer id,
            @RequestBody Mortgage updateMortgage) {

        ApiKeyAuth auth = (ApiKeyAuth) SecurityContextHolder.getContext().getAuthentication();
        User user = (User) auth.getPrincipal();

        Optional<Bank> bankOptional = bankRepository.findById(bankId);
        if (bankOptional.isEmpty()) {
            return ResponseEntity.notFound().build();
        }
        if (!bankOptional.get().ownerId().equals(user.id())) {
            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
        }

        if (!mortgageRepository.existsByIdAndBankId(id, bankId)) {
            return ResponseEntity.notFound().build();
        }
        Mortgage toSave = new Mortgage(id, bankId, updateMortgage.name(), updateMortgage.type(),
                updateMortgage.description(), updateMortgage.TAE());
        return ResponseEntity.ok(mortgageRepository.save(toSave));
    }

    @DeleteMapping("/{id}")
    public ResponseEntity<Mortgage> delete(@PathVariable Integer bankId,
            @PathVariable Integer id) {

        ApiKeyAuth auth = (ApiKeyAuth) SecurityContextHolder.getContext().getAuthentication();
        User user = (User) auth.getPrincipal();

        Optional<Bank> bankOptional = bankRepository.findById(bankId);
        if (bankOptional.isEmpty()) {
            return ResponseEntity.notFound().build();
        }
        if (!bankOptional.get().ownerId().equals(user.id())) {
            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
        }

        if (!mortgageRepository.existsByIdAndBankId(id, bankId)) {
            return ResponseEntity.notFound().build();
        }
        mortgageRepository.deleteById(id);
        return ResponseEntity.noContent().build();
    }

    @GetMapping("/{id}/simulate")
    public ResponseEntity<?> simulate(@PathVariable Integer bankId,
            @PathVariable Integer id,
            @RequestParam double principal,
            @RequestParam int years) {

        try {
            double fee = mortgageSimulationService.simulate(bankId, id, principal, years);
            return ResponseEntity.ok(fee);
        } catch (java.util.NoSuchElementException e) {
            return ResponseEntity.notFound().build();
        } catch (java.lang.IllegalArgumentException e) {
            return ResponseEntity.badRequest().body(e.getMessage());
        } catch (Exception e) {
            return ResponseEntity.internalServerError().body("An error ocurred: " + e.getMessage());
        }
    }

    @GetMapping("/{id}/simulate/pdf")
    public ResponseEntity<?> simulatePdf(@PathVariable Integer bankId,
            @PathVariable Integer id,
            @RequestParam double principal,
            @RequestParam int years) {

        try {
            byte[] pdfBytes = mortgageSimulationService.generateAmortizationPdf(bankId, id, principal, years);

            HttpHeaders headers = new HttpHeaders();
            headers.setContentType(MediaType.APPLICATION_PDF);
            headers.setContentDispositionFormData("attachment", "cuadro_amortizacion_" + id + ".pdf");

            return new ResponseEntity<>(pdfBytes, headers, HttpStatus.OK);

        } catch (java.util.NoSuchElementException e) {
            return ResponseEntity.notFound().build();
        } catch (java.lang.IllegalArgumentException e) {
            return ResponseEntity.badRequest().body(e.getMessage());
        } catch (Exception e) {
            return ResponseEntity.internalServerError().body("An error ocurred: " + e.getMessage());
        }
    }
}